@@ -10,7 +10,7 @@ keywords = ["forecast"]
 license = "MIT"
 license-files = ["LICENSE"]
 authors = [{ name = "monosans", email = "hsyqixco@protonmail.com" }]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 classifiers = [
   "Framework :: Pydantic",
   "Framework :: Pydantic :: 2",
@@ -18,7 +18,6 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
@@ -36,7 +35,7 @@ urls.repository = "https://github.com/monosans/pywttr"
 dev = [
   "covdefaults==2.3",
   "mypy[faster-cache]==1.18.2",
-  "pytest==8.4.2",
+  "pytest==9.0.0",
   "pytest-cov==7.0.0",
   "pytest-github-actions-annotate-failures==0.3",
   "ruff==0.14.4",
